+++ scout/constants/indexes.py
@@ -1,4 +1,4 @@
-from pymongo import (IndexModel, ASCENDING, DESCENDING)
+from pymongo import (IndexModel, ASCENDING, DESCENDING, TEXT)
 
 INDEXES = {
     'hgnc_gene': [
@@ -76,6 +76,13 @@ INDEXES = {
         IndexModel([
             ('description', ASCENDING)],
             name="description"),
+        IndexModel([
+            ('description', TEXT)],
+            default_language='english',
+            name="description_text"),
+        IndexModel([
+            ('hpo_number', ASCENDING)],
+            name="number"),
     ],
     'transcript': [
         IndexModel([
@@ -103,3 +110,4 @@ INDEXES = {
     ],
     
 }
+